+++ src/main/java/com/keyin/airport/AirportRepository.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
+public interface AirportRepository extends JpaRepository<Airport, Long> {
     // Custom query method to find airports by city ID
-    List<Airport> findByCityId(Integer cityId);
+    List<Airport> findByCityId(Long cityId);
 
-    Optional<Airport> findByNameAndCodeAndCityId(String name, String code, Integer cityId);
+    Optional<Airport> findByNameAndCodeAndCityId(String name, String code, Long cityId);
 
 }
