+++ src/main/java/com/keyin/city/City.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class City {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String state;
     private Integer population;
@@ -35,8 +35,8 @@ public class City {
         this.population = population;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
@@ -50,7 +50,7 @@ public class City {
     public List<Airport> getAirports() { return airports; }
     public void setAirports(List<Airport> airports) { this.airports = airports; }
 
-        @Override
+    @Override
     public String toString() {
         return "City{id=" + id + ", name='" + name + "', state='" + state + "', population=" + population + "}";
     }
