+++ src/main/java/com/keyin/passenger/PassengerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PassengerRepository extends JpaRepository<Passenger, Integer> { // Change to Long
+public interface PassengerRepository extends JpaRepository<Passenger, Long> { // Change to Long
 }
